Fix `ClassDefNotFound` & `NullPointerException` errors
